+++ app/Http/Controllers/Auth/Verify/Secondary/Update.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace App\Http\Controllers\Auth\Verify\Email;
+namespace App\Http\Controllers\Auth\Verify\Secondary;
 
 use App\Http\Controllers\Controller;
-use Illuminate\Auth\Events\Verified;
-use Illuminate\Foundation\Auth\EmailVerificationRequest;
+use App\Http\Requests\Auth\Verify\Secondary\Request as SecondaryVerificationRequest;
 use Illuminate\Http\RedirectResponse;
 
 class Update extends Controller
@@ -12,19 +11,19 @@ class Update extends Controller
     /**
      *  Mark the authenticated user's email address as verified.
      *
-     *  @param  \Illuminate\Foundation\Auth\EmailVerificationRequest  $request
+     *  @param  \App\Http\Requests\Auth\Verify\Secondary\Request  $request
      *  @return \Illuminate\Http\RedirectResponse
      */
-    public function __invoke(EmailVerificationRequest $request): RedirectResponse
+    public function __invoke(SecondaryVerificationRequest $request): RedirectResponse
     {
-        if ($request->user()->hasVerifiedEmail()) {
+        if ($request->user()->hasVerifiedSecondary()) {
             return redirect()
                 ->intended($this->route('dashboard').'?verified=1');
         }
 
         $request->fulfill();
 
-        session()->flash('status', 'email-verification-success');
+        session()->flash('status', 'secondary-verification-success');
 
         return redirect()
             ->intended($this->route('dashboard').'?verified=1');