+++ app/Http/Controllers/Users/Destroy.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\ValidationException;
 
 class Destroy extends Controller
 {
@@ -26,6 +27,12 @@ class Destroy extends Controller
         $user = $request->is('my/*')
             ? $request->user()
             : $model::findOrFail($request->user);
+
+        if ($user->is_protected) {
+            throw ValidationException::withMessages([
+                'password' => __('auth.protected'),
+            ]);
+        }
 
         Auth::logout();