+++ app/Models/User.php
@@ -3,7 +3,7 @@
 namespace App\Models;
 
 use Enraiged\Users\Models\User as Authenticatable;
-
+use Illuminate\Auth\Listeners\RegisteredListener;
 class User extends Authenticatable
 {
 }