+++ lang/es/auth.php
@@ -3,7 +3,10 @@
 return [
 
     'failed' => 'Estas credenciales no coinciden con nuestros registros.',
-    'password' => 'La contraseña proporcionada es incorrecta.',
+    'password' => [
+        'La contraseña proporcionada es incorrecta.',
+        'protected' => 'Este usuario está protegido contra cambios de contraseña.',
+    ],
     'protected' => 'Este usuario está protegido y no se puede eliminar.',
     'throttle' => 'Demasiados intentos de inicio de sesión. Inténtelo nuevamente en :seconds segundos.',