+++ lang/fr/auth.php
@@ -3,7 +3,10 @@
 return [
 
     'failed' => 'Ces informations d\'identification ne correspondent pas à nos dossiers.',
-    'password' => 'Le mot de passe fourni est incorrect.',
+    'password' => [
+        'Le mot de passe fourni est incorrect.',
+        'protected' => 'Cet utilisateur est protégé contre les changements de mot de passe.',
+    ],
     'protected' => "Cet utilisateur est protégé et ne peut pas être supprimé",
     'throttle' => 'Trop de tentatives de connexion. Veuillez réessayer dans :seconds secondes.',